Automate sphinx documentation build
This effectively turns this repo into a python project with dependencies
managed by pipenv. Sphinx is used to build the documentation. There's a
github action that runs on every push that just tests to make sure the
docs will build. There's another action that runs on merging to master
that will build the docs and push the changes to the gh-pages branch.