Document how to change default branch
Why these changes are being introduced:

We don't want to use the term 'master' in our GitHub workflows.

Relevant ticket(s):

https://mitlibraries.atlassian.net/browse/ENGX-68

How this addresses that need:

This changes the terminology for the default branch from 'master'
to the currently used 'main', and provides a guide for how to migrate
repos to use 'main' as the default branch.

Side effects of this change:

No immediate side effects other than the recommended changes it
implies. We should also update the default branch in this repo.