Add missing step to GitHub branch renaming guide
Why these changes are being introduced:

In the instructions for renaming the base branch, there is a missing
step to push the renamed branch to the remote repo.

Relevant ticket(s):

N/A

How this addresses that need:

This adds the missing step to the documentation.

Side effects of this change:

N/A