Add Code Climate, Coveralls to branch renaming info
Why these changes are being introduced:

Changing the default branch in GitHub necessitates a corresponding
change in Code Climate. Additionally, there may be some confusion
about Code Climate not running on the first PR. This information is
currently missing from the instructions.

Relevant ticket(s):

N/A

How this addresses that need:

This adds the missing step to the instructions for renaming the
default branch, as well as a note about Coveralls.

Side effects of this change:

I've moved the instructions to what I think is a more logical point
in the document.