Add instructions on how to set up new sites in Matomo
Why these changes are being introduced:

We don't appear to have any local documentation on this process. The
Matomo docs are solid, but it would be useful to document

Relevant ticket(s):

* https://mitlibraries.atlassian.net/browse/ETD-231

How this addresses that need:

This adds some initial documentation on how to add a property to Matomo,
noting a few details about our implementation.

Side effects of this change:

The readme for the docker-matomo repo should be updated to include a
link to these docs (I will make that change shortly).