Update how ENV works in Heroku pipelines
Why these changes are being introduced:

In the new Heroku pipelines, ENV for PR builds is no longer derived from
the staging app, but are instead configured separately.

Relevant ticket(s):

N/A.

How this addresses that need:

This updates the documentation to reflect the new pipeline behavior.

Side effects of this change:

Changed instances of `master` branch to `main`.