AWS and Terraform updates
The instructions in the AWS and Terraform sections were outdated.

The mitlib-terraform repo is fully updated for Terraform 0.12, so
the Tf 0.11 note was removed from the Terraform section. We now manage
all the IAM CLI account creation with Terraform, so the instructions to
configure the key/secret pair for AWS CLI access was updated.