Update Terraform Documentation
Why these changes are being introduced:
Related to our move to an AWS Organization we are also migrating
to a new process for Terraform. First, we are starting to use Terraform
Cloud. Second, we are building all new repositories for our Terraform
code.

How this addresses that need:
* rename the old Terraform page to `legacy-terraform.rst`
* create a new `terraform-1.0.md` page to capture the important new
information about Terraform

Side effects of this change:
None, I hope!