Terraform and GitHub doc updates
Why these changes are being introduced:
This fleshes out some of the details in the new Terraform document. It
also updates the github page with additional information relevant to
Terraform repositories that is a little different than other repos.

How this addresses that need:
* Add appropriate links to the `terraform-1.0.md` document
* Add clarification to the `guthub.md` document with notes specific
to InfraEng and Terraform.

Side effects of this change:
None.