Update instructions to change default branch
Why these changes are being introduced:

When we wrote these docs, GitHub did not provide a way to change
the default branch, but you can how rename the branch easily in
the web UI.

Relevant ticket(s):

N/A.

How this addresses that need:

This updates the corresponding documentation to reference the
easier method.

Side effects of this change:

I'm making some assumptions that Coveralls and commit history in
releases will work using GitHub's method. (Previously, the
number of commits prior to the branch change would be lost, and
Coveralls wouldn't work for the first PR.) I haven't confirmed this,
but it seems unlikely that it will be a concern for many more of our
repos at this point.