Initial documentation on code review
Why these changes are being introduced:

We have a rudimentary guide on pull requests, but we need more
detailed information on the collaborative process of code review.

Relevant ticket(s):

https://mitlibraries.atlassian.net/browse/ENGX-81

How this addresses that need:

This adds a doc to the 'How we work together' section of the guides
that covers the basics of code review: how to prepare a PR, the roles
of the author/reviewer, and some of the nuance inherent in critiquing
someone else's work.

Side effects of this change:

It still feels fairly rudimentary and may need more expansion.