Update workflows to target the correct branch
Why these changes are being introduced:

The deploy and test workflows target `master`, which is no longer
the default branch.

Relevant ticket(s):

N/A.

How this addresses that need:

This updates workflows to target `main`.

Side effects of this change:

Updates links in the `rails` and `vuejs` language docs that go to
nonexistent `master` branches.