Update Matomo config docs
Why these changes are being introduced:

While looking for internal docs on our Matomo implementation's
privacy settings, I realized that much of the info in this doc is
outdated.

Relevant ticket(s):

https://mitlibraries.atlassian.net/browse/UXWS-1532

How this addresses that need:

This removes outdated info (mostly URLs) and removes redundant
info that is either already documented in the UXWS wiki or the
official Matomo docs.

Side effects of this change:

The IP anonymization setting is now documented in the UXWS wiki.