Add unique setup instructions for Matomo
Why these changes are being introduced:

As we add different types of properties to Matomo, we should
update the docs to explain how to set up Matomo in them.

Relevant ticket(s):

https://mitlibraries.atlassian.net/browse/UXWS-1563

How this addresses that need:

This adds instructions on how to set up Matomo in the WP Pantheon
network, LibApps, and static sites. It also stubs out the
instructions for Dome and DSpace, which have not yet been
configured for Matomo.

Side effects of this change:

We'll need to do update the docs again once Dome and DSpace are
unblocked.