Update a11y docs
Why these changes are being introduced:

The Libraries Accessibility Team is working on a LibGuide that
includes a11y guidance for staff. It would be useful to link to
the a11y dev docs in that guide. However, most of the content in this
section was written over six years ago, with a few edits here and there
in the intervening years. Some things have changed, and others could be
expanded upon.

Relevant ticket(s):

Unticketed (part of LAT work mentioned above)

How this addresses that need:

This updates the information in the a11y guide.

Side effects of this change:

There is probably more to add here, but it feels like a decent
starting point.