Update GitHub-related Documentation (#61)
Why these changes are being introduced:
With the updates to user permissions/privileges/roles in our GitHub Organization, it was time to update our developer documentation to reflect these changes. There are also updates to the versions in the GHA workflows to address some errors that cropped up recently.

How this addresses that need:
* Update the GitHub documentation, including the additional information about owners versus members and repository roles
* Fix other slightly-outdated information in other files
* Update the setup-python action to v4 (from v1) in the workflows
* Update the python version to 3.12 (from 3.7) in the workflows
* Update the python version to 3.12 in the Pipfile and updatePipfile.lock